pay重构

+++ mp/user/user.go
@@ -32,13 +32,13 @@ type Info struct {
 	Subscribe     int32    `json:"subscribe"`
 	OpenID        string   `json:"openid"`
 	Nickname      string   `json:"nickname"`
-	Sex           int32    `json:"sex"`
+	Sex           int      `json:"sex"`
 	City          string   `json:"city"`
 	Country       string   `json:"country"`
 	Province      string   `json:"province"`
 	Language      string   `json:"language"`
 	Headimgurl    string   `json:"headimgurl"`
-	SubscribeTime int32    `json:"subscribe_time"`
+	SubscribeTime int64    `json:"subscribe_time"`
 	UnionID       string   `json:"unionid"`
 	Remark        string   `json:"remark"`
 	GroupID       int32    `json:"groupid"`
